requirements.txt
aiogram==3.13.1
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
babel==2.17.0
certifi==2025.8.3
Django==5.2.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
environs==14.3.0
frozenlist==1.7.0
gunicorn==21.2.0
idna==3.10
marshmallow==4.0.1
multidict==6.6.4
pillow==11.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
redis==5.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
yarl==1.20.1
uvicorn==0.34.0
httpx==0.27.2
dj-database-url==2.2.0
celery==5.4.0
django-celery-results==2.5.1
python-json-logger==2.0.7
django-cors-headers==4.3.1
